+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TablesSidebarDMDController.java
@@ -126,7 +126,7 @@ public class TablesSidebarDMDController implements Initializable {
   @FXML
   private void onDMDPosition() {
     GameRepresentation g = game.get();
-    TableDialogs.openDMDPositionDialog(g, null);
+    TableDialogs.openDMDPositionDialog(g, tablesSidebarController.getTableOverviewController());
   }
 
   @FXML