+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TablesSidebarDirectB2SController.java
@@ -219,7 +219,7 @@ public class TablesSidebarDirectB2SController implements Initializable, StudioEv
   @FXML
   private void onDMDPosition() {
     if (game.isPresent()) {
-      TableDialogs.openDMDPositionDialog(game.get(), null);
+      TableDialogs.openDMDPositionDialog(game.get(), tablesController.getTableOverviewController());
     }
   }