+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TablesSidebarMediaController.java
@@ -328,7 +328,7 @@ public class TablesSidebarMediaController implements Initializable {
   @FXML
   private void onDMDPosition() {
     if (game.isPresent()) {
-      TableDialogs.openDMDPositionDialog(game.get(), null);
+      TableDialogs.openDMDPositionDialog(game.get(), tablesSidebarController.getTableOverviewController());
     }
   }