+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/panels/BaseGameModel.java
@@ -1,0 +1,7 @@
+package de.mephisto.vpin.ui.tables.panels;
+
+public interface BaseGameModel {
+
+  public int getGameId();
+
+}